# docs/api/openapi.yaml
openapi: 3.0.3
info:
  title: Runic API
  description: |
    Firewall rule management system API.
    
    ## Authentication
    Most endpoints require Bearer token authentication. Include the token in the Authorization header:
    ```
    Authorization: Bearer <token>
    ```
    
    ## Error Responses
    All endpoints may return error responses in the following format:
    ```json
    {
      "error": "error message"
    }
    ```
    
    Common HTTP status codes:
    - 400: Bad Request - Invalid input
    - 401: Unauthorized - Missing or invalid token
    - 403: Forbidden - Insufficient permissions
    - 404: Not Found - Resource not found
    - 409: Conflict - Resource in use
    - 500: Internal Server Error
  version: 1.0.0
  contact:
    name: Runic Support
  license:
    name: MIT
servers:
  - url: http://localhost:8080/api/v1
    description: Local development server
  - url: /api/v1
    description: Production server (relative path)
...
paths:
...
  /alerts:
    get:
      tags:
      - Alerts
      summary: List all alerts
      description: Returns alert history. Requires admin role.
      operationId: listAlerts
      responses:
        '200':
          description: Successful response
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Alert'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '500':
          $ref: '#/components/responses/InternalServerError'
      security:
      - BearerAuth: []
    delete:
      tags:
      - Alerts
      summary: Clear all alerts
      description: Deletes all alert history. Requires admin role.
      operationId: clearAllAlerts
      responses:
        '200':
          description: Alerts cleared
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: "alerts cleared"
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '500':
          $ref: '#/components/responses/InternalServerError'
      security:
      - BearerAuth: []
  /alerts/{id}:
    get:
      tags:
      - Alerts
      summary: Get alert by ID
      description: Returns a specific alert. Requires admin role.
      operationId: getAlert
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: integer
          format: int64
      responses:
        '200':
          description: Successful response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Alert'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/InternalServerError'
      security:
      - BearerAuth: []
    delete:
      tags:
      - Alerts
      summary: Delete alert
      description: Deletes a specific alert. Requires admin role.
      operationId: deleteAlert
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: integer
          format: int64
      responses:
        '200':
          description: Alert deleted
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: "alert deleted"
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/InternalServerError'
      security:
      - BearerAuth: []
  /alert-rules:
    get:
      tags:
      - Alerts
      summary: List alert rules
      description: Returns all alert rules. Requires admin role.
      operationId: listAlertRules
      responses:
        '200':
          description: Successful response
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/AlertRule'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '500':
          $ref: '#/components/responses/InternalServerError'
      security:
      - BearerAuth: []
  /alert-rules/{id}:
    put:
      tags:
      - Alerts
      summary: Update alert rule
      description: Updates an existing alert rule. Requires admin role.
      operationId: updateAlertRule
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: integer
          format: int64
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AlertRule'
      responses:
        '200':
          description: Alert rule updated
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: "alert rule updated"
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/InternalServerError'
      security:
      - BearerAuth: []
  /settings/smtp:
    get:
      tags:
      - Alerts
      summary: Get SMTP configuration
      description: Returns the SMTP configuration for sending alerts. Requires admin role.
      operationId: getSMTPConfig
      responses:
        '200':
          description: Successful response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SMTPConfig'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '500':
          $ref: '#/components/responses/InternalServerError'
      security:
      - BearerAuth: []
    put:
      tags:
      - Alerts
      summary: Update SMTP configuration
      description: Updates the SMTP configuration. Requires admin role.
      operationId: updateSMTPConfig
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/SMTPConfig'
      responses:
        '200':
          description: SMTP configuration updated
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: "smtp updated"
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '500':
          $ref: '#/components/responses/InternalServerError'
      security:
      - BearerAuth: []
  /settings/smtp/test:
    post:
      tags:
      - Alerts
      summary: Test SMTP configuration
      description: Sends a test email using the current SMTP configuration. Requires admin role.
      operationId: testSMTP
      responses:
        '200':
          description: Test email sent
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: "test email sent"
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '500':
          $ref: '#/components/responses/InternalServerError'
      security:
      - BearerAuth: []
  /users/me/notification-preferences:
    get:
      tags:
      - Alerts
      summary: Get notification preferences
      description: Returns the current user's notification preferences.
      operationId: getNotificationPrefs
      responses:
        '200':
          description: Successful response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/NotificationPreferences'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '500':
          $ref: '#/components/responses/InternalServerError'
      security:
      - BearerAuth: []
    put:
      tags:
      - Alerts
      summary: Update notification preferences
      description: Updates the current user's notification preferences.
      operationId: updateNotificationPrefs
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/NotificationPreferences'
      responses:
        '200':
          description: Preferences updated
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: "preferences updated"
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '500':
          $ref: '#/components/responses/InternalServerError'
      security:
      - BearerAuth: []

  # ==================== IMPORTS ====================
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      description: Bearer token authentication
  schemas:
...
    Alert:
      type: object
      description: Alias for AlertHistory — represents a triggered alert record
      properties:
        id:
          type: integer
          format: uint
        rule_id:
          type: integer
          format: uint
        alert_type:
          type: string
          enum: [peer_offline, bundle_failed, blocked_spike, peer_online, new_peer, bundle_deployed]
        peer_id:
          type: integer
          nullable: true
        peer_hostname:
          type: string
        severity:
          type: string
          enum: [info, warning, critical]
        subject:
          type: string
        message:
          type: string
        metadata:
          type: string
          description: JSON string for additional context
        status:
          type: string
          enum: [pending, sent, failed, throttled]
        sent_at:
          type: string
          format: date-time
          nullable: true
        error_message:
          type: string
          nullable: true
        created_at:
          type: string
          format: date-time
    AlertRule:
      type: object
      properties:
        id:
          type: integer
          format: uint
        name:
          type: string
          maxLength: 255
        alert_type:
          type: string
          enum: [peer_offline, bundle_failed, blocked_spike, peer_online, new_peer, bundle_deployed]
        enabled:
          type: boolean
        threshold_value:
          type: integer
          description: Threshold value (e.g., blocked count for spike)
        threshold_window_minutes:
          type: integer
          description: Time window for threshold evaluation
        peer_id:
          type: integer
          nullable: true
          description: Peer ID if rule applies to a specific peer (null = all peers)
        throttle_minutes:
          type: integer
          description: Minimum time between alerts for this rule
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
...
    SMTPConfig:
      type: object
      properties:
        host:
          type: string
          maxLength: 255
        port:
          type: integer
          minimum: 1
          maximum: 65535
        username:
          type: string
          maxLength: 255
        use_tls:
          type: boolean
          default: true
        from_address:
          type: string
          format: email
          maxLength: 255
        enabled:
          type: boolean
          default: false
...
    NotificationPreferences:
      type: object
      properties:
        id:
          type: integer
          format: uint
        user_id:
          type: integer
          format: uint
        enabled_alerts:
          type: string
          description: JSON array of alert types (empty means all enabled)
        quiet_hours_enabled:
          type: boolean
          default: false
        quiet_hours_start:
          type: string
          description: Start time in HH:MM format
          maxLength: 5
        quiet_hours_end:
          type: string
          description: End time in HH:MM format
          maxLength: 5
        quiet_hours_timezone:
          type: string
          maxLength: 50
        digest_enabled:
          type: boolean
          default: false
        digest_frequency:
          type: string
          enum: [daily, weekly]
          default: daily
        digest_time:
          type: string
          description: Time in HH:MM format
          maxLength: 5
          default: "09:00"
        digest_timezone:
          type: string
          maxLength: 50
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
...
    Error:
      type: object
      properties:
        error:
          type: string
  responses:
    BadRequest:
      description: Bad Request - Invalid input
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
    Unauthorized:
      description: Unauthorized - Missing or invalid token
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
    Forbidden:
      description: Forbidden - Insufficient permissions
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
    NotFound:
      description: Not Found - Resource not found
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
...
    InternalServerError:
      description: Internal Server Error
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'